// src/types/types.d.ts
export interface components {
  schemas: {
// ...
    WineListType: {
      /** Format: double */
      id: number;
      name: string;
      region: string;
      image: string;
      /** Format: double */
      price: number;
      type: string;
      /** Format: double */
      avgRating: number;
      /** Format: double */
      reviewCount: number;
      recentReview: {
        user: components["schemas"]["ReviewUser"];
        /** Format: date-time */
        updatedAt: string;
        /** Format: date-time */
        createdAt: string;
        content: string;
        aroma: components["schemas"]["Aroma"][];
        /** Format: double */
        rating: number;
        /** Format: double */
        id: number;
      } | null;
      /** Format: double */
      userId: number;
    };
// ...
    WineDetailType: {
      /** Format: double */
      id: number;
      name: string;
      region: string;
      image: string;
      /** Format: double */
      price: number;
      type: string;
      /** Format: double */
      avgRating: number;
      /** Format: double */
      reviewCount: number;
      recentReview: {
        user: components["schemas"]["ReviewUser"];
        /** Format: date-time */
        updatedAt: string;
        /** Format: date-time */
        createdAt: string;
        content: string;
        aroma: components["schemas"]["Aroma"][];
        /** Format: double */
        rating: number;
        /** Format: double */
        id: number;
      } | null;
      /** Format: double */
      userId: number;
      reviews: components["schemas"]["ReviewListType"][];
      avgRatings: components["schemas"]["Record_number.number_"];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}